+++ Source/Epiphany.Model/Entity/ReadStatusFeedItemModel.cs
@@ -17,7 +17,7 @@ namespace Epiphany.Model
             }
         }
 
-        protected override int GetId(GoodreadsUpdate update)
+        protected override long GetId(GoodreadsUpdate update)
         {
             return update.Object.ReadStatus.Id;
         }
