+++ Source/Epiphany.Model/Entity/UserChallengeFeedItemModel.cs
@@ -26,7 +26,7 @@ namespace Epiphany.Model
             }
         }
 
-        protected override int GetId(GoodreadsUpdate update)
+        protected override long GetId(GoodreadsUpdate update)
         {
             return 0;
         }
